lambda_functions/auth/requirements.txt
boto3==1.28.56
botocore==1.31.56
jmespath==1.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.16
